mcp_weather_service/pyproject.toml
[project]
name = "mcp_weather_service"
version = "0.1.0"
description = "A simple MCP weather server example."
authors = [{ name = "AI Assistant", email = "assistant@example.com" }]
dependencies = [
    "mcp",
    "httpx",
    "python-dotenv",
    "pydantic",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "respx",
]

[project.scripts]
run-server = "mcp_weather_service:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest",
    "pytest-asyncio",
    "respx",
]

[tool.hatch.metadata]
allow-direct-references = true 